[ci] Extract shared queryRewardCouponsV2 into DbTransferInputQueries
