+++ apps/common/src/main/scala/org/lfdecentralizedtrust/splice/store/db/DbTransferInputQueries.scala
@@ -4,12 +4,13 @@
 package org.lfdecentralizedtrust.splice.store.db
 
 import com.daml.ledger.javaapi.data.codegen.ContractId
+import org.lfdecentralizedtrust.splice.codegen.java.splice.amulet.RewardCouponV2
 import org.lfdecentralizedtrust.splice.codegen.java.splice.round.IssuingMiningRound
 import org.lfdecentralizedtrust.splice.codegen.java.splice.types.Round
 import org.lfdecentralizedtrust.splice.store.MultiDomainAcsStore.ContractCompanion
 import org.lfdecentralizedtrust.splice.store.db.AcsQueries.{AcsStoreId, SelectFromAcsTableResult}
 import org.lfdecentralizedtrust.splice.store.{Limit, LimitHelpers, TransferInputStore}
-import org.lfdecentralizedtrust.splice.util.{Contract, TemplateJsonDecoder}
+import org.lfdecentralizedtrust.splice.util.{Contract, ContractWithState, TemplateJsonDecoder}
 import org.lfdecentralizedtrust.splice.util.FutureUnlessShutdownUtil.futureUnlessShutdownToFuture
 import com.digitalasset.canton.lifecycle.CloseContext
 import com.digitalasset.canton.resource.DbStorage
@@ -84,5 +85,33 @@ trait DbTransferInputQueries extends AcsQueries with AcsTables with LimitHelpers
             contract -> issuance
         }
     }
+  }
+
+  protected def queryRewardCouponsV2(
+      includeUnassigned: Boolean,
+      includeAssigned: Boolean,
+      limit: Limit,
+  )(implicit tc: TraceContext): Future[Seq[
+    ContractWithState[RewardCouponV2.ContractId, RewardCouponV2]
+  ]] = {
+    val whereClause = (includeUnassigned, includeAssigned) match {
+      case (true, true) => sql""
+      case (true, false) => sql"and acs.create_arguments->>'beneficiary' is null"
+      case (false, true) => sql"and acs.create_arguments->>'beneficiary' is not null"
+      case (false, false) => sql"and false"
+    }
+    for {
+      result <- dbStorage.query(
+        selectFromAcsTableWithState(
+          acsTableName,
+          acsStoreId,
+          domainMigrationId,
+          RewardCouponV2.COMPANION,
+          additionalWhere = whereClause,
+          orderLimit = sql"order by acs.contract_expires_at asc limit ${sqlLimit(limit)}",
+        ),
+        "listRewardCouponsV2",
+      )
+    } yield result.map(contractWithStateFromRow(RewardCouponV2.COMPANION)(_))
   }
 }
