+++ apps/wallet/src/main/scala/org/lfdecentralizedtrust/splice/wallet/store/db/DbExternalPartyWalletStore.scala
@@ -22,7 +22,6 @@ import org.lfdecentralizedtrust.splice.wallet.store.ExternalPartyWalletStore
 import com.digitalasset.canton.lifecycle.CloseContext
 import com.digitalasset.canton.logging.NamedLoggerFactory
 import com.digitalasset.canton.resource.DbStorage
-import slick.jdbc.canton.ActionBasedSQLInterpolation.Implicits.actionBasedSQLInterpolationCanton
 import com.digitalasset.canton.tracing.TraceContext
 import com.digitalasset.canton.util.ShowUtil.*
 import com.digitalasset.canton.topology.ParticipantId
@@ -77,7 +76,6 @@ class DbExternalPartyWalletStore(
     with LimitHelpers {
 
   import org.lfdecentralizedtrust.splice.store.db.AcsQueries.AcsStoreId
-  import org.lfdecentralizedtrust.splice.util.FutureUnlessShutdownUtil.futureUnlessShutdownToFuture
   import multiDomainAcsStore.waitUntilAcsIngested
 
   override protected def acsStoreId: AcsStoreId = multiDomainAcsStore.acsStoreId
@@ -100,26 +98,7 @@ class DbExternalPartyWalletStore(
     ContractWithState[amuletCodegen.RewardCouponV2.ContractId, amuletCodegen.RewardCouponV2]
   ]] =
     waitUntilAcsIngested {
-      val whereClause = (includeUnassigned, includeAssigned) match {
-        case (true, true) => sql""
-        case (true, false) => sql"and acs.create_arguments->>'beneficiary' is null"
-        case (false, true) => sql"and acs.create_arguments->>'beneficiary' is not null"
-        case (false, false) => sql"and false"
-      }
-      for {
-        result <- storage.query(
-          selectFromAcsTableWithState(
-            WalletTables.externalPartyAcsTableName,
-            acsStoreId,
-            domainMigrationId,
-            amuletCodegen.RewardCouponV2.COMPANION,
-            additionalWhere = whereClause,
-            orderLimit = sql"order by acs.contract_expires_at asc limit ${sqlLimit(limit)}",
-          ),
-          "listRewardCouponsV2",
-        )
-      } yield result
-        .map(contractWithStateFromRow(amuletCodegen.RewardCouponV2.COMPANION)(_))
+      queryRewardCouponsV2(includeUnassigned, includeAssigned, limit)
     }
 
   override def listSortedLivenessActivityRecords(
